Battle System v 0.5f
Proto Battle system having a turn by turn system. Currently in a free form manner. 
Basic AI and Pathfinding .

Expansion of various method to better link the Editor and the logic.

Cleaning up some files
Log at chokepoint to see what's going on under the hood . The game can be play in a console.

Animations and In game Actor for better performance.